base: Actually fix the padding on StatusBar
Android P gets a new QS layout that looks good
but it has some issues on smaller screens devices,
like the rounded buttons clipping with the borders
and the date not looking really good.
LineageOS suggested a fix (https://review.lineageos.org/224021)
but it's circumventing it because it adds padding
 and some people doesn't like the padding, so, instead
let's apply the 8dp spacing over the affected headers
and the small qs tile.

Change-Id: I4c01945bfb3d9f58f10cd7deb392a2d40bc50289
Signed-off-by: VenkatVishalV <venkatvishal124@gmail.com>
Signed-off-by: sayan7848 <sayan7848@gmail.com>